% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{articles_capitalised}
\alias{articles_capitalised}
\title{Catalan articles capitalised (including apostrophed)}
\format{
An object of class \code{character} of length 11.
}
\usage{
articles_capitalised
}
\description{
Catalan articles capitalised (including apostrophed)
}
\keyword{datasets}
