% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{articles_lowercase}
\alias{articles_lowercase}
\title{Catalan articles in lowercase}
\format{
An object of class \code{character} of length 11.
}
\usage{
articles_lowercase
}
\description{
Catalan articles in lowercase
}
\keyword{datasets}
