% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{catmunalias}
\alias{catmunalias}
\title{Catalan municipality names}
\format{
Data frame with columns
\describe{
\item{alias}{Nome used for a municipality}
\item{mun}{Cannonical name for the municipality}
}
}
\source{
<https://www.idescat.cat/dev/api/emex/>

<https://ca.wikipedia.org/wiki/Llista_de_municipis_de_Catalunya>
}
\usage{
catmunalias
}
\description{
DB of different ways of refering to catalan municipalities.
}
\examples{
  catmunalias
}
\keyword{datasets}
