% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{check_injective_mapping}
\alias{check_injective_mapping}
\title{Given a mapping xp -> yp, check if all elements in x are present in x}
\usage{
check_injective_mapping(x, xp, yp)
}
\arguments{
\item{x}{origin set to find}

\item{xp}{origin set in reference data}

\item{yp}{target set in reference data}
}
\value{
no value. Stops if failing.
}
\description{
Given a mapping xp -> yp, check if all elements in x are present in x
}
