% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affixes.R
\name{from_prefix_to_suffix}
\alias{from_prefix_to_suffix}
\title{Returns the list of words, replacing the prefix by the corresponding suffix,
if matching. If it doesn't match, it results the word as is.}
\usage{
from_prefix_to_suffix(words, prefixes, suffixes)
}
\arguments{
\item{words}{words that will be transformed}

\item{prefixes}{list of prefixes to search for}

\item{suffixes}{corresponsing suffixes that will be prepended if a prefix
matches}
}
\value{
A list of the transformed words
}
\description{
Returns the list of words, replacing the prefix by the corresponding suffix,
if matching. If it doesn't match, it results the word as is.
}
