% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affixes.R
\name{from_suffix_to_prefix}
\alias{from_suffix_to_prefix}
\title{Returns the list of words, replacing the suffix by the corresponding prefix,
if matching. If it doesn't match, it results the word as is.}
\usage{
from_suffix_to_prefix(words, suffixes, prefixes)
}
\arguments{
\item{words}{words that will be transformed}

\item{suffixes}{list of suffixes to search for}

\item{prefixes}{corresponsing prefixes that will be prepended if a suffix
matches}
}
\value{
A list of the transformed words
}
\description{
Returns the list of words, replacing the suffix by the corresponding prefix,
if matching. If it doesn't match, it results the word as is.
}
