% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aliases.R
\name{get_unmapped_aliases}
\alias{get_unmapped_aliases}
\title{Get all elements in x not found in xp}
\usage{
get_unmapped_aliases(x, xp)
}
\arguments{
\item{x}{origin set to find}

\item{xp}{origin set in reference data}
}
\value{
xnotfound = xp - x
}
\description{
Get all elements in x not found in xp
}
