% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affixes.R
\name{matchFirst}
\alias{matchFirst}
\title{Returns with element from 'from' matches 'words' using 'compare_function'
Returns Inf if none}
\usage{
matchFirst(words, from, compare_function)
}
\arguments{
\item{words}{words that will be compared}

\item{from}{elements to compare to}

\item{compare_function}{Function to use in comparison}
}
\value{
A vector indexes of the same length of words, with positions in from. NA if none matched for that word.
}
\description{
Returns with element from 'from' matches 'words' using 'compare_function'
Returns Inf if none
}
