% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{prefix_capitalised}
\alias{prefix_capitalised}
\title{Catalan articles capitalised (includes space if not apostrophed)}
\format{
An object of class \code{character} of length 11.
}
\usage{
prefix_capitalised
}
\description{
Catalan articles capitalised (includes space if not apostrophed)
}
\keyword{datasets}
