% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{prefix_lowercase}
\alias{prefix_lowercase}
\title{Catalan articles lowercase (includes space if not apostrophed)}
\format{
An object of class \code{character} of length 11.
}
\usage{
prefix_lowercase
}
\description{
Catalan articles lowercase (includes space if not apostrophed)
}
\keyword{datasets}
