% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{suffix_capitalised}
\alias{suffix_capitalised}
\title{Catalan articles capitalised as suffix, separated by comma}
\format{
An object of class \code{character} of length 11.
}
\usage{
suffix_capitalised
}
\description{
Catalan articles capitalised as suffix, separated by comma
}
\keyword{datasets}
