% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/articles.R
\docType{data}
\name{suffix_lowercase}
\alias{suffix_lowercase}
\title{Catalan articles in lowercase, separated by comma}
\format{
An object of class \code{character} of length 11.
}
\usage{
suffix_lowercase
}
\description{
Catalan articles in lowercase, separated by comma
}
\keyword{datasets}
