% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catmunalias.R
\docType{data}
\name{catmunalias}
\alias{catmunalias}
\title{Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1340 rows and 2 columns.
}
\usage{
catmunalias
}
\description{
Dataset
}
\keyword{datasets}
